+++ packages/data-types/src/index.ts
@@ -79,6 +79,8 @@ export * from './lib/interfaces/enemy.interface';
 export * from './lib/interfaces/chance-per-item.interface';
 export * from './lib/interfaces/custom-entry.interface';
 export * from './lib/interfaces/product-size-by-mood.interface';
+export * from './lib/interfaces/animal-data.interface';
+export * from './lib/interfaces/animal-shop-data.interface';
 
 
 export * from './lib/enums/quality.enum';