+++ packages/guide/src/app/locations/locations.module.ts
@@ -25,7 +25,7 @@ import { MerfolkOracleTailStoreComponent } from './components/shops/merfolk-orac
 import { PetShopComponent } from './components/pet-shop/pet-shop.component';
 import { PetAdoptionDetailsComponent } from './components/pet-adoption-details/pet-adoption-details.component';
 import { MatTooltipModule } from '@angular/material/tooltip';
-import { RanchComponent } from './components/shops/ranch/ranch.component';
+import { RanchComponent } from './components/ranch/ranch.component';
 import { BeachShackComponent } from './components/shops/beach-shack/beach-shack.component';
 import { ConcernedMonkeyComponent } from './components/shops/concerned-monkey/concerned-monkey.component';
 import { MeritShopComponent } from './components/merit-shop/merit-shop.component';
@@ -33,6 +33,7 @@ import { MeritShopDetailsComponent } from './components/merit-shop-details/merit
 import { BosComponent } from './components/shops/bos/bos.component';
 import { BensCaravanComponent } from './components/shops/bens-caravan/bens-caravan.component';
 import { SocketAndPanComponent } from './components/shops/socket-and-pan/socket-and-pan.component';
+import { AnimalDetailsComponent } from './components/animal-details/animal-details.component';
 
 @NgModule({
     declarations: [
@@ -62,6 +63,7 @@ import { SocketAndPanComponent } from './components/shops/socket-and-pan/socket-
         BosComponent,
         BensCaravanComponent,
         SocketAndPanComponent,
+        AnimalDetailsComponent,
     ],
     imports: [
         CommonModule,