+++ packages/guide/src/app/shared/components/requirements-list/requirements-list.component.ts
@@ -1,0 +1,10 @@
+import { Component, Input } from '@angular/core';
+import { RequirementEntry } from "@ci/data-types";
+
+@Component({
+    selector: 'app-requirements-list',
+    templateUrl: './requirements-list.component.html',
+})
+export class RequirementsListComponent {
+    @Input() requirements?: RequirementEntry
+}