+++ packages/pak-files-parser/src/types/datatable-ref.type.ts
@@ -1,0 +1,7 @@
+export type DatatableRef = {
+    "DataTable": {
+        "ObjectName": string
+        "ObjectPath": string
+    },
+    "RowName": "None" | string
+}